Clean up context awareness mocks
