+++ src/platform/plugins/shared/discover/public/context_awareness/__mocks__/profile_provider_services.ts
@@ -1,0 +1,25 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import {
+  createLogsContextServiceMock,
+  createTracesContextServiceMock,
+} from '@kbn/discover-utils/src/__mocks__';
+import { createDiscoverServicesMock } from '../../__mocks__/services';
+import type { ProfileProviderServices } from '../profile_providers/profile_provider_services';
+
+export const createProfileProviderServicesMock = (
+  discoverServices = createDiscoverServicesMock()
+): ProfileProviderServices => {
+  return {
+    ...discoverServices,
+    logsContextService: createLogsContextServiceMock(),
+    tracesContextService: createTracesContextServiceMock(),
+  };
+};
